GPF-1606 CNVLoader now uses flexible variant loader
