+++ dae/dae/backends/cnv/tests/test_load_cnv_ng.py
@@ -5,7 +5,8 @@ import io
 from dae.genomic_resources.test_tools import convert_to_tab_separated
 from dae.pedigrees.loader import FamiliesLoader
 
-from dae.backends.cnv.loader import CNVLoader
+from dae.backends.cnv.loader import _configure_cnv_location, \
+    _configure_cnv_best_state
 
 
 @pytest.fixture
@@ -42,7 +43,7 @@ def test_configure_cnv_loader_vcf_like_pos():
     header = ["person_id", "Chr", "Start", "End", "variant", "extra"]
     transformers = []
 
-    CNVLoader._configure_cnv_location(
+    _configure_cnv_location(
         header, transformers,
         cnv_chrom="Chr",
         cnv_start="Start",
@@ -59,7 +60,7 @@ def test_configure_cnv_loader_vcf_like_pos_missmatch():
     transformers = []
 
     with pytest.raises(ValueError):
-        CNVLoader._configure_cnv_location(
+        _configure_cnv_location(
             header, transformers,
             cnv_chrom="Chrom",
             cnv_start="Start",
@@ -70,7 +71,7 @@ def test_configure_cnv_loader_vcf_like_pos_missmatch():
 def test_configure_cnv_loader_location():
     header = ["person_id", "location", "variant", "extra"]
     transformers = []
-    CNVLoader._configure_cnv_location(header, transformers)
+    _configure_cnv_location(header, transformers)
     assert header == [
         "person_id", "location", "variant", "extra"]
 
@@ -79,7 +80,7 @@ def test_configure_cnv_loader_location_mismatch():
     header = ["person_id", "location", "variant", "extra"]
     transformers = []
     with pytest.raises(ValueError):
-        CNVLoader._configure_cnv_location(
+        _configure_cnv_location(
             header, transformers, cnv_location="Location")
 
 
@@ -89,7 +90,7 @@ def test_configure_cnv_loader_dae_best_state_default(
     header = ["family_id", "location", "variant", "best_state", "extra"]
     transformers = []
 
-    CNVLoader._configure_cnv_best_state(
+    (
         header, transformers,
         families, gpf_instance_2013.reference_genome,
     )
@@ -104,7 +105,7 @@ def test_configure_cnv_loader_dae_best_state(
     header = ["familyId", "location", "variant", "bestState", "extra"]
     transformers = []
 
-    CNVLoader._configure_cnv_best_state(
+    _configure_cnv_best_state(
         header, transformers,
         families, gpf_instance_2013.reference_genome,
         cnv_family_id="familyId",
@@ -121,7 +122,7 @@ def test_configure_cnv_loader_person_id(
     header = ["personId", "location", "variant", "extra"]
     transformers = []
 
-    CNVLoader._configure_cnv_best_state(
+    _configure_cnv_best_state(
         header, transformers,
         families, gpf_instance_2013.reference_genome,
         cnv_person_id="personId",
@@ -138,7 +139,7 @@ def test_configure_cnv_loader_best_state_mixed(
     transformers = []
 
     with pytest.raises(ValueError):
-        CNVLoader._configure_cnv_best_state(
+        _configure_cnv_best_state(
             header, transformers,
             families, gpf_instance_2013.reference_genome,
             cnv_person_id="personId",
@@ -153,7 +154,7 @@ def test_configure_cnv_loader_best_state_person_id_not_found(
     transformers = []
 
     with pytest.raises(ValueError):
-        CNVLoader._configure_cnv_best_state(
+        _configure_cnv_best_state(
             header, transformers,
             families, gpf_instance_2013.reference_genome,
             cnv_person_id="person_id",
@@ -167,7 +168,7 @@ def test_configure_cnv_loader_best_state_best_state_not_found(
     transformers = []
 
     with pytest.raises(ValueError):
-        CNVLoader._configure_cnv_best_state(
+        _configure_cnv_best_state(
             header, transformers,
             families, gpf_instance_2013.reference_genome,
             cnv_family_id="familyId",
@@ -182,7 +183,7 @@ def test_configure_cnv_loader_best_state_family_id_not_found(
     transformers = []
 
     with pytest.raises(ValueError):
-        CNVLoader._configure_cnv_best_state(
+        _configure_cnv_best_state(
             header, transformers,
             families, gpf_instance_2013.reference_genome,
             cnv_family_id="family_id",
